refactor(services): move service interfaces to contracts package
Relocated `IPlayerService` and `CrudService` interfaces to the `contracts` subpackage for better organization and clarity. Updated the package declarations to reflect this change.
