+++ src/main/java/com/javierdesant/spring_sport_flow/infrastructure/services/contracts/IPlayerService.java
@@ -1,4 +1,4 @@
-package com.javierdesant.spring_sport_flow.infrastructure.services;
+package com.javierdesant.spring_sport_flow.infrastructure.services.contracts;
 
 import com.javierdesant.spring_sport_flow.api.dto.requests.PlayerRequest;
 import com.javierdesant.spring_sport_flow.api.dto.responses.PlayerResponse;
